+++ CHANGELOG.md
@@ -13,3 +13,7 @@ Initial Release
 - Direct solving using Gaussian Elimination with Partial Pivoting for Dense Matrices
 - Matrix inverses calculated using Gaussian Elimination with Partial Pivoting for Dense Matrices
 - Iterative solvers for Dense and Sparse Matrices, using the Conjugate Gradient Method.
+
+## Version 1.0.1
+- Bug Fix (DenseMatrix.solve/DenseMatrix.inverse): pivot selection error when a column contains no positive numbers.
+- Bug Fix: Conversion from DenseMatrix to SparseMatrix now correctly works with negatives.